feat: Designed 3 pages for this app.
Developed the UI for pages, incorporating previously created components to keep code minimal. React icons were installed and used in a number of intances, register and login buttons were linked to the pages they redirect to using Links from react-router-dom, app icon was updated to a colour that blends better with the app's theme.