frontend form validation and integration of endpoints to fetch data from backend while logging in and registering
Token was saved in browser's local storage, services to send form data to the backend in order to register a new user and for their logging in to the main app. Installed react-toastify for successful login and signin notifications. Toaster container was added to app.js. The user is redirected to landing page once they successfully login, this was accomplished with useNavigate. The token stored locally was decoded with jwt-decode for its payload, and finally, token is deleted from the local storage when a user logout.