feat: Moved EditModal from the NotesContainer component and created DeleteModal. Successfully passed states from parent component NotesContainer to children components DeleteModal, EditModal and NoteCard. A couple of small fixes were also incorporated.
Separate modals that will be opened depending on which of the 2 buttons on the NoteCard component is clicked, were created. The EditModal will allow user to edit or update the content and title of the chosen note while DeleteModal will allow the preview of a note before finally discarding it from the list. Moved the states of the selected card out of NoteCard into parent NoteContainer component to allow the 3 children components inherit states from here. The value of the states were altered as needed from the children components by using call back of state setter functions. Also added onClick actions to the edit and delete buttons of each NoteCard to trigger opening of the desired modal. Finally, as a fix, renamed the route for the main page of app from '/home' to '/notes' in app.js.