feat: Completed the functionality of NoteReader page to enable update and deletion of a single note retrieved based on its _id, i.e. noteId. Additional minor fixes were also done on previous pages/components
A number of components were modified to enable the full functionality of NoteReader page. In NoteContainer.jsx, noteId was saved to local storage so as to access it for note deletion and note update functionality from within NoteReader page. React Link to redirect user to NoteReader page added to body of note in NoteCard. An additional Confirmation component was added to improve user experience by giving the user an option for extra confirmation before carrying out an irreversible deletion operation. Created functions that call the services which edit and delete individual note in the NoteReader component. Finally, removed unused useNavigate from the LandingPage component.